Address review: move proc paths into platform.ts, socket = ipc.sock
- Fold the per-subprocess path helpers into util/platform.ts, next to the
  existing managed-process path helpers (worker PID files, monitoring
  dir), instead of a standalone proc-paths.ts module. Delete proc-paths.ts.
- Name the IPC socket `ipc.sock` (fixed) rather than `<name>.sock` — the
  procs/<name>/ directory already carries the subprocess name.
- Trim the now-redundant procs-convention section from routes/AGENTS.md
  down to a pointer at the platform.ts helpers.
